Updated tailwind helpers and config
Setup separate file for tailwind helper JS
Added logic for opt-in importing for color palettes in blocks (includes dedup to prevent overwrites)

I tested this locally using a custom block style.
